bilinear resize and rotate

+++ bilinear/checker_props/checker-bilinear_resize.m
@@ -4,7 +4,7 @@
 %! y = dlmread("../tests/in/factory1.txt");
 %! resized = bilinear_resize(img, x, y);
 %! resized_ref = dlmread("../tests/ref/bilinear/resized1.txt");
-%! assert(max(max(abs(resized - resized_ref))) <= 5);
+%! assert(max(max(abs(int32(resized) - resized_ref))) <= 5);
 
 %!test
 %! img = imread("../tests/in/sample2.png");
@@ -12,7 +12,7 @@
 %! y = dlmread("../tests/in/factory2.txt");
 %! resized = bilinear_resize(img, x, y);
 %! resized_ref = dlmread("../tests/ref/bilinear/resized2.txt");
-%! assert(max(max(abs(resized - resized_ref))) <= 5);
+%! assert(max(max(abs(int32(resized) - resized_ref))) <= 5);
 
 %!test
 %! img = imread("../tests/in/sample2.png");
@@ -20,7 +20,7 @@
 %! y = dlmread("../tests/in/factory3.txt");
 %! resized = bilinear_resize(img, x, y);
 %! resized_ref = dlmread("../tests/ref/bilinear/resized3.txt");
-%! assert(max(max(abs(resized - resized_ref))) <= 5);
+%! assert(max(max(abs(int32(resized) - resized_ref))) <= 5);
 
 %!test
 %! img = imread("../tests/in/sample3.png");
@@ -28,7 +28,7 @@
 %! y = dlmread("../tests/in/factory4.txt");
 %! resized = bilinear_resize(img, x, y);
 %! resized_ref = dlmread("../tests/ref/bilinear/resized4.txt");
-%! assert(max(max(abs(resized - resized_ref))) <= 5);
+%! assert(max(max(abs(int32(resized) - resized_ref))) <= 5);
 
 %!test
 %! img = imread("../tests/in/sample3.png");
